backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.1",
    "bcrypt==4.1.2",
    "cryptography==41.0.7",
    "fastapi==0.109.0",
    "fastapi-cors==0.0.6",
    "httpx==0.26.0",
    "imbalanced-learn==0.11.0",
    "ipykernel==6.29.0",
    "ipywidgets==8.1.1",
    "joblib==1.3.2",
    "jupyter==1.0.0",
    "kaleido==0.2.1",
    "lightgbm==4.1.0",
    "loguru==0.7.2",
    "matplotlib==3.8.2",
    "nbconvert==7.14.2",
    "numpy==1.26.2",
    "openpyxl==3.1.2",
    "pandas==2.1.4",
    "passlib[bcrypt]==1.7.4",
    "plotly==5.18.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "email-validator==2.1.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "pytz==2023.3",
    "scikit-learn==1.3.2",
    "scipy==1.11.4",
    "seaborn==0.13.0",
    "sqlalchemy==2.0.25",
    "tqdm==4.66.1",
    "uvicorn[standard]==0.27.0",
    "xgboost==2.0.3",
]
